requirements.txt
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
Flask
Flask-Session
Flask-cors
